+++ tests/Fixtures/App/Kernel.php
@@ -32,7 +32,7 @@ class Kernel extends HttpKernel
         }
     }
 
-    public function registerContainerConfiguration(LoaderInterface $loader)
+    public function registerContainerConfiguration(LoaderInterface $loader): void
     {
         $loader->load(__DIR__.'/config/config.php');
     }